refactor : Carousel api 디렉토리를 hooks에서 modules 로 변경
